requirements.txt
flask
pymongo
line-bot-sdk
jieba
python-dotenv
matplotlib 
requests